Split documentation into small Sphinx-Doc projects
From each DocBook (it's reST conversion) a small, stand-alone Sphinx-Doc
project is generated. *Autodoc* from source code and the *misc-docs* are
also each one stand-alone project. All projects these are referred from
the *mainpage* project.

A Makefile with all the reST and HTML targets was added. This is also a
preparation to publish on github-pages, see issue 1:

* https://github.com/return42/sphkerneldoc/issues/1