DocBook2reST: set kernel_doc_mode = "kernel-doc"
Set parser's default kernel-doc mode to "kernel-doc" for migrated
content.

Since "reST" is the default mode of the kernel-doc parser and the
migrated DocBook content is based on *vintage* kernel-doc comments, it
is the best to run the parser in the *vintage* mode, as long as we
consume old, unrevised kernel-doc markups.

Signed-off-by: Markus Heiser <markus.heiser@darmarIT.de>